refactor ClusterEventLog -> directory
