+++ src/components/diagnostic/DiagnosticPane.js
@@ -3,7 +3,7 @@ import 'semantic-ui-css/semantic.min.css';
 import { Grid } from 'semantic-ui-react';
 import './DiagnosticPane.css';
 import GeneratePackage from './GeneratePackage';
-import ClusterEventLog from './ClusterEventLog';
+import ClusterEventLog from './ClusterEventLog/ClusterEventLog';
 import Ping from './Ping';
 import uuid from 'uuid';
 import Explainer from '../ui/scaffold/Explainer/Explainer';
